Split param resolution into its own module
